requirements.txt
google-adk>=0.0.1
mcp[cli]>=1.1.2
fastmcp>=0.4.1
google-genai>=0.3.0
